SourceCode/requirements.txt
flask
langchain
pandas
tiktoken
huggingface_hub
python-dotenv
# Additional libraries for specific functionalities
numpy       
scikit-learn
scipy

